traffic lights can have individual durations,
