+++ intersection.py
@@ -4,31 +4,32 @@ import time
 import random
 import threading
 from direction import Direction
-stop_sign_delay = 2
+from constants import *
 
 # represents types of intersections that can exist in the traffic simulation
 class Intersection:
     pass
 
 class StopLight(Intersection):
-    
-    def __init__(self):
+    def __init__(self, duration):
         self.y_axis_green = bool(random.randint(0,1))
         self.queue = [] # (index of car, direction)
+        self.duration = duration
+        self.release_delay = 1
 
     def flip_light(self, ts):
         self.y_axis_green = not self.y_axis_green
-        # print('queue', self.queue)
         if len(self.queue) > 0:
             self.process_queue(ts)
+        threading.Timer(self.duration, self.flip_light, kwargs={'ts': ts}).start()
     
     def process_queue(self, ts):    
         car_index, direction = self.queue.pop(0)  # Get first car in queue
-        ts.move_car_to_next_intersection(car_index, direction)
+        ts.release_car_from_queue(car_index, direction)
         
         # If there are more cars in queue, schedule the next one
         if len(self.queue) > 0:
-            threading.Timer(1.0, self.process_queue, args=[ts]).start()
+            threading.Timer(LIGHT_RELEASE_DELAY, self.process_queue, args=[ts]).start()
 
     def join_queue(self, car_index, direction, ts):
         if (car_index, direction) in self.queue:
@@ -41,7 +42,7 @@ class StopLight(Intersection):
             self.queue.append((car_index,direction))
             # print('attempting to move at a red light')
             return
-        ts.move_car_to_next_intersection(car_index, direction)
+        ts.release_car_from_queue(car_index, direction)
     
 class StopSign(Intersection):
     def __init__(self):
@@ -53,15 +54,14 @@ class StopSign(Intersection):
             print('car already at intersection')
             return
         
-        if self.time_since_last_pass + stop_sign_delay < time.time():
-            ts.move_car_to_next_intersection(car_index, direction)
+        if self.time_since_last_pass + STOP_SIGN_DELAY < time.time():
+            ts.release_car_from_queue(car_index, direction)
             self.time_since_last_pass = time.time()
         else:
             self.queue.append(car_index)
-            move_after = max(0, stop_sign_delay - (time.time()- self.time_since_last_pass))
-            # print(move_after
+            move_after = max(0, STOP_SIGN_DELAY - (time.time()- self.time_since_last_pass))
             threading.Timer(move_after, self.move_car_thread, kwargs={'car_index': car_index, 'direction': direction, 'ts': ts}).start()
 
     def move_car_thread(self, car_index, direction, ts):
         self.queue.remove(car_index)
-        ts.move_car_to_next_intersection(car_index, direction)
+        ts.release_car_from_queue(car_index, direction)
